--- src/features/paper-list/apis/index.ts
@@ -1,2 +1,0 @@
-export * from "./get-recipients"
-export * from "./types"
