--- src/features/paper-list/models/index.ts
@@ -1,2 +1,0 @@
-export { usePopularRecipients } from "./use-popular-recipients"
-export { useRecentRecipients } from "./use-recent-recipients"
