+++ src/features/paper-list/models/use-popular-recipients.ts
@@ -2,9 +2,9 @@ import { useSuspenseQuery } from "@tanstack/react-query"
 
 import { recipientQueryKey, RECIPIENTS_LIMIT, RECIPIENTS_OFFSET } from "@/entities/recipient"
 
-import { getRecipients } from "../apis"
+import { getRecipients } from "../apis/get-recipients"
 
-import type { RecipientsParams } from "../apis"
+import type { RecipientsParams } from "../apis/types"
 import type { Recipient } from "@/entities/recipient"
 
 export function usePopularRecipients({ limit = RECIPIENTS_LIMIT, offset = RECIPIENTS_OFFSET }: RecipientsParams) {
